python_processing/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.15.0
pillow==10.1.0
scikit-learn==1.3.2
psycopg2-binary==2.9.9
boto3==1.29.7
python-dotenv==1.0.0
watchdog==3.0.0
werkzeug==3.0.1
